bot/requirements.txt
httpx==0.28.1
requests==2.32.3
schedule==1.2.2
python-dotenv==1.1.0
praw==7.8.1
pyyaml==6.0.2
